+++ scilab/routines/scicos/relay.c
@@ -14,7 +14,7 @@ double *inptr[],*outptr[],*t;
     double *y;
     double *u;
     int nev,ic;
-    ic=z[0];
+    ic=(int)z[0];
     if ((*flag)<3) {
       if ((*nevprt)>0) {
 	ic=-1;
